+++ Frontend/src/app/modules/gcode/sites/gcode-edit/gcode-edit.component.ts
@@ -12,6 +12,7 @@ import { AppState } from '../../../../store/app.state';
 import { Settings } from '../../../shared/components/settings/settings.component';
 import { SnackbarService } from '../../../../services/snackbar.service';
 import { LoadingService } from '../../../shared/services/loading.service';
+import { GcodeFunctionsService } from '../../services/gcode-functions.service';
 
 export interface GalleryEntryUpload {
   name?: string;
@@ -33,7 +34,8 @@ export class GcodeEditComponent implements OnInit, AfterViewInit {
     private store: Store,
     private router: Router,
     private snackbarService: SnackbarService,
-    private loadingService: LoadingService
+    private loadingService: LoadingService,
+    private gcodeFunctions: GcodeFunctionsService
   ) {}
 
   @ViewChild(CanvasGcodeRendererComponent) renderer:
@@ -118,14 +120,20 @@ export class GcodeEditComponent implements OnInit, AfterViewInit {
 
     console.log(this.gcodeViewerService.scaleToDrawingArea);
     if (this.gcodeViewerService.gcodeType != 'custom') {
-      gcodeArray = this.replacePenCommands(gcodeArray);
+      gcodeArray = this.gcodeFunctions.replacePenCommands(
+        gcodeArray,
+        this.settings.penDownCommand,
+        this.settings.penUpCommand
+      );
       if (this.gcodeViewerService.scaleToDrawingArea) {
-        gcodeArray = this.scaleGcode(
+        gcodeArray = this.gcodeFunctions.scaleGcode(
           gcodeArray,
-          this.settings.centerOnDrawingArea
+          this.settings.centerOnDrawingArea,
+          this.settings.selectedPaperProfile.drawingOffset,
+          this.settings.selectedPaperProfile.paperMax
         );
       }
-      gcodeArray = this.applyOffset(
+      gcodeArray = this.gcodeFunctions.applyOffset(
         gcodeArray,
         this.settings.selectedPaperProfile.drawingOffset
       );
@@ -154,153 +162,4 @@ export class GcodeEditComponent implements OnInit, AfterViewInit {
     this.loadingService.isLoading = false;
     this.router.navigate(['gcode', 'drawing']);
   }
-
-  replacePenCommands(gcode: string[]): string[] {
-    for (let i = 0; i < gcode.length; i++) {
-      if (gcode[i].includes('M03') || gcode[i].includes('M3')) {
-        gcode[i] = this.settings.penDownCommand;
-      } else if (gcode[i].includes('M05') || gcode[i].includes('M5')) {
-        gcode[i] = this.settings.penUpCommand;
-      }
-    }
-    return gcode;
-  }
-
-  scaleGcode(gcode: string[], center: Boolean): string[] {
-    let gcodeScaling = 1;
-    let biggest: number[] = [0, 0];
-    let centeringOffset: number[] = [0, 0];
-
-    gcode = this.moveToOrigin(gcode);
-
-    for (let i = 0; i < gcode.length; i++) {
-      let cmd = this.getG1Parameter(gcode[i]);
-      if (biggest[0] < cmd[0]) {
-        biggest[0] = cmd[0];
-      }
-      if (biggest[1] < cmd[1]) {
-        biggest[1] = cmd[1];
-      }
-    }
-
-    let scalings: number[] = [
-      (this.settings.selectedPaperProfile.paperMax[0] -
-        this.settings.selectedPaperProfile.drawingOffset[0]) /
-        biggest[0],
-      (this.settings.selectedPaperProfile.paperMax[1] -
-        this.settings.selectedPaperProfile.drawingOffset[1]) /
-        biggest[1],
-    ];
-
-    if (scalings[0] < scalings[1]) {
-      gcodeScaling = scalings[0];
-      if (center) {
-        centeringOffset[1] =
-          (this.settings.selectedPaperProfile.paperMax[1] -
-            this.settings.selectedPaperProfile.drawingOffset[1] -
-            biggest[1] * gcodeScaling) /
-          2;
-      }
-    } else {
-      gcodeScaling = scalings[1];
-      if (center) {
-        centeringOffset[0] =
-          (this.settings.selectedPaperProfile.paperMax[0] -
-            this.settings.selectedPaperProfile.drawingOffset[0] -
-            biggest[0] * gcodeScaling) /
-          2;
-      }
-    }
-
-    for (let i = 0; i < gcode.length; i++) {
-      let command = gcode[i];
-      if (
-        command.startsWith('G1') ||
-        command.startsWith('G01') ||
-        command.startsWith('G0')
-      ) {
-        let parameter = this.getG1Parameter(command);
-        parameter[0] = parameter[0] * gcodeScaling + centeringOffset[0];
-        parameter[1] = parameter[1] * gcodeScaling + centeringOffset[1];
-
-        //round floating points
-        parameter[0] = this.round(parameter[0], this.settings.floatingPoints);
-        parameter[1] = this.round(parameter[1], this.settings.floatingPoints);
-        gcode[i] = 'G1 X' + parameter[0] + 'Y' + parameter[1];
-      }
-    }
-    return gcode;
-  }
-
-  //moves the gcode to the origing by substracting the smallest x,y value from all x,y values
-  moveToOrigin(gcode: string[]): string[] {
-    let smallest: number[] = [-1, -1];
-
-    for (let i = 0; i < gcode.length; i++) {
-      if (gcode[i].startsWith('G1')) {
-        let cmd = this.getG1Parameter(gcode[i]);
-        if (smallest[0] == -1) {
-          smallest[0] = cmd[0];
-          smallest[1] = cmd[1];
-        } else {
-          if (smallest[0] > cmd[0]) {
-            smallest[0] = cmd[0];
-          }
-          if (smallest[1] > cmd[1]) {
-            smallest[1] = cmd[1];
-          }
-        }
-      }
-    }
-
-    for (let i = 0; i < gcode.length; i++) {
-      let command = gcode[i];
-      if (command.startsWith('G1')) {
-        let parameter = this.getG1Parameter(command);
-        parameter[0] = parameter[0] - smallest[0];
-        parameter[1] = parameter[1] - smallest[1];
-
-        gcode[i] = 'G1 X' + parameter[0] + 'Y' + parameter[1];
-      }
-    }
-
-    return gcode;
-  }
-
-  applyOffset(gcodeArray: string[], offset: number[]): string[] {
-    for (let i = 0; i < gcodeArray.length; i++) {
-      let command = gcodeArray[i];
-      if (command.startsWith('G1')) {
-        let parameter = this.getG1Parameter(command);
-        parameter[0] += offset[0];
-        parameter[1] += offset[1];
-
-        //round floating points
-        parameter[0] = this.round(parameter[0], this.settings.floatingPoints);
-        parameter[1] = this.round(parameter[1], this.settings.floatingPoints);
-
-        gcodeArray[i] = 'G1 X' + parameter[0] + 'Y' + parameter[1];
-      }
-    }
-
-    return gcodeArray;
-  }
-
-  getG1Parameter(command: string): number[] {
-    let x: number = parseFloat(
-      command
-        .substring(command.indexOf('X') + 1, command.lastIndexOf('Y'))
-        .trim()
-    );
-    let y: number = parseFloat(
-      command.substring(command.indexOf('Y') + 1, command.length).trim()
-    );
-    return [x, y];
-  }
-
-  // from: https://stackoverflow.com/questions/11832914/how-to-round-to-at-most-2-decimal-places-if-necessary
-  private round = (n: number, dp: number) => {
-    const h = +'1'.padEnd(dp + 1, '0'); // 10 or 100 or 1000 or etc
-    return Math.round(n * h) / h;
-  };
 }
