+++ Frontend/src/app/modules/shared/components/terminal-button/terminal-button.component.scss
@@ -1,5 +1,5 @@
 .terminal-button {
-  position: absolute;
+  position: fixed;
   width: 56px;
   height: 56px;
   border: 3px solid rgb(0, 0, 0);
